requirements.txt
alembic==1.15.1
aniso8601==10.0.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==4.0.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
dnspython==2.7.0
email-validator==2.0.0
et_xmlfile==2.0.0
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
fonttools==4.56.0
greenlet==3.1.1
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.2
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
numpy==1.24.3
openpyxl==3.1.2
oscrypto==1.3.0
pandas==1.5.3
pdfkit==1.0.0
Pillow==10.0.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
pypdf==5.4.0
pyphen==0.17.2
pypng==0.20220715.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.0.4
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.20
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
Werkzeug==2.3.7
wkhtmltopdf==0.2
WTForms==3.0.1
xhtml2pdf==0.2.17
xlrd==2.0.1
XlsxWriter==3.2.2
zopfli==0.2.3.post1
gunicorn==20.1.0
psycopg2-binary==2.9.9
setuptools==68.0.0

